fixing battle simulator

+++ Game/Setup/UnitFactory.cs
@@ -54,8 +54,8 @@ namespace Game.Setup {
                                                                 (ArmorClass)
                                                                 Enum.Parse(typeof(ArmorClass), toks[col["ArmrClass"]].ToUpper()),
                                                                 ushort.Parse(toks[col["Hp"]]),
-                                                                byte.Parse(toks[col["Atk"]]),
-                                                                byte.Parse(toks[col["Def"]]),
+                                                                ushort.Parse(toks[col["Atk"]]),
+                                                                ushort.Parse(toks[col["Def"]]),
                                                                 byte.Parse(toks[col["Rng"]]),
                                                                 byte.Parse(toks[col["Stl"]]),
                                                                 byte.Parse(toks[col["Spd"]]),
